+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
+- Refactor `aem_saml` module to use `swaqger_aem_osgi` API Client for setting SAML configuration
 - Upgrade ruby_aem to 3.6.0
 
 ## [5.5.0] - 2020-04-05